Add files via upload
the second project is to build a tool that assesses the strength of a password based on criteria such as length, presence of uppercase and lowercse letters, numbers, and pecial characters, provide feedback to user on the password strength. 

The program started by importing the re module which provides support for working with regular expression, which are used to search for patterns in strings.

The def function check_password_strength, takes one argument pwd which is the password to be checked.

Then the criteria to determine a strong password was listed which includes the following
1.	Password should be at least 8 characters long
2.	Password should have at least one uppercase letter
3.	Password should have at least one lowercase letter
4.	Password should contain at least one numbers
5.	Password should contain at least one special character 

The next stage sums up the criteria and determined the state of the password depending on the number of criteria met. Such as 
1.	If all criteria are met, the password is “very strong”
2.	If 4 criteria are met, the password is “strong”
3.	If 3 criteria are met, the password is “moderate”
4.	If 2 criteria are met, the password is “weak”
5.	If fewer than 2 criteria are met, the password is “very weak”

The next function provides the feedback string that starts with a password strength. If one or more of the criteria is not met, it suggests adding the missing criteria to make the password very strong. The user is prompted to input a password while the feedback will be printed to know the strength of the password as shown in the screenshots below.

This program can check the strengths of passwords and provides feedback to the users by suggesting ways to improve the strength of their passwords. 

According to NIST (National institute for standard and technology) and other professional bodies, a password capable of resisting cracking should meet the following criteria.
1.	It should be at least minimum of 8 characters and maximum of 12-16 characters long.
2.	It should have at least one uppercase and lowercase letter, at least one numbers and one special character. 
Additionally, the use of common words, phrase or easily guessable information like date of births and telephone numbers should be avoided. If possible, a password manger should be used to generate and store complex passwords.
The more complex the password is, the higher the amount of time needed to crack the password.